Revert "It is safer to use "uintptr_t" when we transfer "void *".And I use"
The commit message garbage, the whole description is on a single line
and it's missing Signed-off-by line.

This reverts commit 69f89fbd4caea0a3e72daa0ee567c11b1ec723fe.

+++ testcases/kernel/syscalls/brk/brk01.c
@@ -39,8 +39,6 @@
 #include <signal.h>
 #include <sys/param.h>
 #include <sys/resource.h>
-#include <stdint.h>
-#include <inttypes.h>
 
 #include "test.h"
 
@@ -56,8 +54,8 @@ void cleanup();
 char *TCID = "brk01";
 int TST_TOTAL = 1;
 
-uintptr_t Max_brk_byte_size;
-uintptr_t Beg_brk_val;
+long Max_brk_byte_size;
+long Beg_brk_val;
 
 #if !defined(UCLINUX)
 
@@ -65,9 +63,9 @@ int main(int ac, char **av)
 {
 	int lc;
 	int incr;
-	uintptr_t nbrkpt;		/* new brk point value */
-	uintptr_t cur_brk_val;	/* current size returned by sbrk */
-	uintptr_t aft_brk_val;	/* current size returned by sbrk */
+	long nbrkpt;		/* new brk point value */
+	long cur_brk_val;	/* current size returned by sbrk */
+	long aft_brk_val;	/* current size returned by sbrk */
 
 	tst_parse_opts(ac, av, NULL, NULL);
 
@@ -92,7 +90,7 @@ int main(int ac, char **av)
 		 * every odd lc value, strink by one incr.
 		 * If lc is equal to 3, no change, special case.
 		 */
-		cur_brk_val = (uintptr_t)sbrk(0);
+		cur_brk_val = (long)sbrk(0);
 		if (lc == 3) {
 			nbrkpt = cur_brk_val;	/* no change, special one time case */
 		} else if ((lc % 2) == 0) {
@@ -120,21 +118,21 @@ int main(int ac, char **av)
 
 		if (TEST_RETURN == -1) {
 
-			aft_brk_val = (uintptr_t)sbrk(0);
+			aft_brk_val = (long)sbrk(0);
 			tst_resm(TFAIL | TTERRNO,
-				 "brk(%"PRIuPTR") failed (size before %"PRIuPTR", after %"PRIuPTR")",
+				 "brk(%ld) failed (size before %ld, after %ld)",
 				 nbrkpt, cur_brk_val, aft_brk_val);
 
 		} else {
-			aft_brk_val = (uintptr_t)sbrk(0);
+			aft_brk_val = (long)sbrk(0);
 			if (aft_brk_val == nbrkpt) {
 
 				tst_resm(TPASS,
-					 "brk(%"PRIuPTR") returned %"PRIuPTR", new size verified by sbrk",
+					 "brk(%ld) returned %ld, new size verified by sbrk",
 					 nbrkpt, TEST_RETURN);
 			} else {
 				tst_resm(TFAIL,
-					 "brk(%"PRIuPTR") returned %"PRIuPTR", sbrk before %"PRIuPTR", after %"PRIuPTR"",
+					 "brk(%ld) returned %ld, sbrk before %ld, after %ld",
 					 nbrkpt, TEST_RETURN,
 					 cur_brk_val, aft_brk_val);
 			}
@@ -189,7 +187,7 @@ void setup(void)
 	if (max_size > (usr_mem_sz / 4))
 		max_size = usr_mem_sz / 4;	/* only fourth mem by single test */
 
-	Beg_brk_val = (uintptr_t)sbrk(0);
+	Beg_brk_val = (long)sbrk(0);
 
 	/*
 	 * allow at least 4 times a big as current.
